+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - completions for bash, zsh, fish, and powershell
 - formula for `brew` package manager
 - global config support
+- formatted output formats usinf `--format` flag
+	- JSON
+	- Pretty JSON
 
 ### Changed
 - ignore paths are now entired handled by todor bin